pyproject.toml
[tool.poetry]
name = "f1-ml"
version = "0.1.0"
description = "End-to-end ML system for F1 race predictions"
authors = ["Francesco Parra <francescoparra02a@gmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.13"

fastf1 = "^3.3.9"
pandas = "^2.1.4"
numpy = "^1.26.2"

scikit-learn = "^1.3.2"
xgboost = "^2.0.3"
lightgbm = "^4.1.0"

matplotlib = "^3.8.2"
seaborn = "^0.13.0"
mlflow = "^2.9.2"

# Utils
python-dotenv = "^1.0.0"
tqdm = "^4.66.1"
scipy-stubs = "~=1.16.3"
pandas-stubs = "~=2.3.3"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
pytest = "^7.4.3"
ipdb = "^0.13.13"

[tool.poetry.group.notebooks.dependencies]
jupyterlab = "^4.0.9"
notebook = "^7.0.6"
ipykernel = "^6.27.1"

[tool.poetry.group.test.dependencies]
flake8 = "^6.0.0"
mypy = "^1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
